requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.4
pandas==2.2.2
joblib==1.4.2
PyJWT==2.8.0
tradelocker==0.56.2
polars==1.15.0
numpy==2.1.1
pytest==8.3.3
gymnasium==0.29.1

